// src/types.ts
export type UserRole = 'warga' | 'pengurus';

export interface UserAccount {
  id: string;
  username: string;
  role: UserRole;
  passwordHash: string;
  namaLengkap: string;
  noKK?: string;
}
// ...
export interface ToastMessage {
  id: string;
  type: 'success' | 'error' | 'info';
  message: string;
}